+++ src/parser/inline/reference.test.ts
@@ -49,19 +49,19 @@ describe('Unit: parser/inline/reference', () => {
       assert.deepStrictEqual(inspect(parser('[[[[a]]]]')), [['<sup class="reference">[[a]]</sup>'], '']);
     });
 
-    it('alias', () => {
-      assert.deepStrictEqual(inspect(parser('[[^a]]')), [['<sup class="reference" data-alias="a"></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a b]]')), [['<sup class="reference" data-alias="a b"></sup>'], '']);
+    it('abbr', () => {
+      assert.deepStrictEqual(inspect(parser('[[^a]]')), [['<sup class="reference" data-abbr="a"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a b]]')), [['<sup class="reference" data-abbr="a b"></sup>'], '']);
       assert.deepStrictEqual(inspect(parser('[[^a  b]]')), [['<sup class="reference">^a  b</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a|]]')), [['<sup class="reference" data-alias="a"></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a|b]]')), [['<sup class="reference" data-alias="a">b</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a|b ]]')), [['<sup class="reference" data-alias="a">b </sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a| ]]')), [['<sup class="reference" data-alias="a"></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a| b]]')), [['<sup class="reference" data-alias="a">b</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^a| b ]]')), [['<sup class="reference" data-alias="a">b </sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^Xyz2020]]')), [['<sup class="reference" data-alias="Xyz2020"></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^X, 2020, p1-2]]')), [['<sup class="reference" data-alias="X, 2020, p1-2"></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[^X. Y., Z et al., 2020, p1-2]]')), [['<sup class="reference" data-alias="X. Y., Z et al., 2020, p1-2"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a|]]')), [['<sup class="reference" data-abbr="a"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a|b]]')), [['<sup class="reference" data-abbr="a">b</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a|b ]]')), [['<sup class="reference" data-abbr="a">b </sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a| ]]')), [['<sup class="reference" data-abbr="a"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a| b]]')), [['<sup class="reference" data-abbr="a">b</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^a| b ]]')), [['<sup class="reference" data-abbr="a">b </sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^Xyz2020]]')), [['<sup class="reference" data-abbr="Xyz2020"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^X, 2020, p1-2]]')), [['<sup class="reference" data-abbr="X, 2020, p1-2"></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[^X. Y., Z et al., 2020, p1-2]]')), [['<sup class="reference" data-abbr="X. Y., Z et al., 2020, p1-2"></sup>'], '']);
     });
 
   });